# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MotoRentR API
  version: 1.0.0
  description: |
    Full-stack motorcycle rental system API for MotoRentR.
    Includes health check, motorcycles listing, rental booking, and basic metadata.
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
...
  /api/motorcycles:
    get:
      summary: List motorcycles (catalog)
      description: >
        Browse the motorcycle catalog. Does not check availability — use
        /api/motorcycles/availability for that. `availability` and `filter_info`
        are always null in the response.
      parameters:
        - in: query
          name: price_category
          schema:
            type: string
            enum: [type-ss, type-s, type-a, type-b, type-c, type-minibike]
          required: false
          description: Price category of motorcycle.
        - in: query
          name: moto_type
          schema:
            type: string
            enum: [cruiser, naked, sport, superbike]
          required: false
          description: Type/category of motorcycle.
        - in: query
          name: brand
          schema:
            type: string
          required: false
          description: Brand name (e.g. KAWASAKI, YAMAHA, HONDA).
      responses:
        '200':
          description: Motorcycles list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MotorcyclesListResponse'
              example:
                count: 2
                data:
                  - id: 1
                    image: /KAWASAKI_NINJA_400.png
                    title: KAWASAKI NINJA 400
                    brand: KAWASAKI
                    price_category: type-s
                    price: 5000
                    moto_type: sport
                    engine_displacement: 399cc
                    max_horsepower: 45hp
                    max_torque: 38Nm
                    engine_type: 水冷四行程並列雙汽缸 DOHC 4V
                    fuel_tank_capacity: 14L
                    seat_height: 785mm
                    weight: 168kg
                    availability: null
                    filter_info: null
                filters_applied:
                  price_category: type-s
                  moto_type: null
                  brand: null
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Database error
  /api/motorcycles/availability:
    get:
      summary: Search motorcycle availability
      description: >
        Search motorcycles with real-time availability for a branch/date/time/duration.
        branch, date, start_time, and duration are all required together — this endpoint
        always performs the conflict check (no silent skip).
      parameters:
        - in: query
          name: branch
          schema:
            type: string
            enum: [taipei, taichung, tainan]
          required: true
          description: Branch of rental shop.
        - in: query
          name: date
          schema:
            type: string
            format: date
          required: true
          description: Rental date (YYYY-MM-DD).
        - in: query
          name: start_time
          schema:
            type: string
            pattern: '^[0-2][0-9]:[0-5][0-9]$'
          required: true
          description: Start time (HH:MM, 24h).
        - in: query
          name: duration
          schema:
            type: string
            enum: [5h, 10h, 24h, 48h]
          required: true
          description: Rental duration.
        - in: query
          name: price_category
          schema:
            type: string
            enum: [type-ss, type-s, type-a, type-b, type-c, type-minibike]
          required: false
          description: Price category of motorcycle.
        - in: query
          name: moto_type
          schema:
            type: string
            enum: [cruiser, naked, sport, superbike]
          required: false
          description: Type/category of motorcycle.
        - in: query
          name: brand
          schema:
            type: string
          required: false
          description: Brand name (e.g. KAWASAKI, YAMAHA, HONDA).
      responses:
        '200':
          description: Motorcycles list with availability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MotorcyclesListResponse'
              example:
                count: 2
                data:
                  - id: 1
                    image: /KAWASAKI_NINJA_400.png
                    title: KAWASAKI NINJA 400
                    brand: KAWASAKI
                    price_category: type-s
                    price: 5000
                    moto_type: sport
                    engine_displacement: 399cc
                    max_horsepower: 45hp
                    max_torque: 38Nm
                    engine_type: 水冷四行程並列雙汽缸 DOHC 4V
                    fuel_tank_capacity: 14L
                    seat_height: 785mm
                    weight: 168kg
                    availability:
                      available: true
                      message: Available for rent
                    filter_info:
                      branch: taipei
                      date: 2025-12-15
                      start_time: 14:00
                      duration: 24h
                filters_applied:
                  branch: taipei
                  date: 2025-12-15
                  start_time: 14:00
                  duration: 24h
                  price_category: type-s
                  moto_type: null
                  brand: null
        '400':
          description: Missing or invalid required parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: branch is required
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Database error
  /api/motorcycles/{motorcycle_id}:
    get:
      summary: Get motorcycle details
      description: Get detailed information for a single motorcycle, including prices for all durations.
      parameters:
        - in: path
          name: motorcycle_id
          required: true
          schema:
            type: integer
          description: Motorcycle ID.
      responses:
        '200':
          description: Motorcycle details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MotorcycleWithPrices'
              example:
                data:
                  id: 1
                  image: /KAWASAKI_NINJA_400.png
                  title: KAWASAKI NINJA 400
                  brand: KAWASAKI
                  price_category: type-s
                  moto_type: sport
                  engine_displacement: 399cc
                  max_horsepower: 45hp
                  max_torque: 38Nm
                  engine_type: 水冷四行程並列雙汽缸 DOHC 4V
                  fuel_tank_capacity: 14L
                  seat_height: 785mm
                  weight: 168kg
                  prices:
                    '5h': 2500
                    '10h': 4000
                    '24h': 5000
                    '48h': 9000
        '404':
          description: Motorcycle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Motorcycle not found
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Database error
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MotorcycleAvailability:
      type: object
      properties:
        available:
          type: boolean
        message:
          type: string
    MotorcycleFilterInfo:
      type: object
      properties:
        branch:
          type: string
          enum: [taipei, taichung, tainan]
        date:
          type: string
          format: date
        start_time:
          type: string
          nullable: true
          pattern: '^[0-2][0-9]:[0-5][0-9]$'
        duration:
          type: string
          enum: [5h, 10h, 24h, 48h]
    Motorcycle:
      type: object
      properties:
        id:
          type: integer
        image:
          type: string
        title:
          type: string
        brand:
          type: string
        price_category:
          type: string
          enum: [type-ss, type-s, type-a, type-b, type-c, type-minibike]
        price:
          type: number
          format: float
          description: Price for selected duration (if applicable).
        moto_type:
          type: string
          enum: [cruiser, naked, sport, superbike]
        engine_displacement:
          type: string
        max_horsepower:
          type: string
        max_torque:
          type: string
        engine_type:
          type: string
          description: Descriptive text (can be in Chinese).
        fuel_tank_capacity:
          type: string
        seat_height:
          type: string
        weight:
          type: string
        availability:
          $ref: '#/components/schemas/MotorcycleAvailability'
        filter_info:
          $ref: '#/components/schemas/MotorcycleFilterInfo'
    MotorcycleWithPrices:
      type: object
      properties:
        id:
          type: integer
        image:
          type: string
        title:
          type: string
        brand:
          type: string
        price_category:
          type: string
          enum: [type-ss, type-s, type-a, type-b, type-c, type-minibike]
        moto_type:
          type: string
          enum: [cruiser, naked, sport, superbike]
        engine_displacement:
          type: string
        max_horsepower:
          type: string
        max_torque:
          type: string
        engine_type:
          type: string
        fuel_tank_capacity:
          type: string
        seat_height:
          type: string
        weight:
          type: string
        prices:
          type: object
          properties:
            '5h':
              type: number
            '10h':
              type: number
            '24h':
              type: number
            '48h':
              type: number
    MotorcyclesListResponse:
      type: object
      properties:
        count:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/Motorcycle'
        filters_applied:
          type: object
          properties:
            branch:
              type: string
              nullable: true
              enum: [taipei, taichung, tainan, null]
            date:
              type: string
              format: date
              nullable: true
            start_time:
              type: string
              nullable: true
            duration:
              type: string
              nullable: true
              enum: [5h, 10h, 24h, 48h, null]
            price_category:
              type: string
              nullable: true
              enum: [type-ss, type-s, type-a, type-b, type-c, type-minibike, null]
            moto_type:
              type: string
              nullable: true
              enum: [cruiser, naked, sport, superbike, null]
            brand:
              type: string
              nullable: true
...